refactor: Organizar archivos de carpetas
